requirements.txt
torch
diffusers
transformers
accelerate
librosa
imageio[ffmpeg]
requests
python-dotenv
tomli
Pillow